feat: no memory copy during ORT inference + refactoring
